Changed live TPS to be averages over a few ( 150 ) transaction

+++ src/main/java/com/sstr/load/manager/Job.java
@@ -8,7 +8,7 @@ public class Job implements Runnable {
 
     private final Scenario scenario;
     private boolean active = true;
-    private long liveTPS = 0;
+    private double liveTPS = 0;
 
 
     public Job(Scenario scenario) {
@@ -19,7 +19,7 @@ public class Job implements Runnable {
         active = false;
     }
 
-    public long getLiveTPS() {
+    public double getLiveTPS() {
         return liveTPS;
     }
